requirements.txt
requests
python-dotenv
brcrypt